Fixed empty definition bug.

+++ modules/template/models/TemplateContentActiveRecord.php
@@ -115,7 +115,8 @@ abstract class TemplateContentActiveRecord extends ActiveRecord
         if($this->isDefinitionContent() && $definition->validate() && $definition->hasValues()) {
             $definition->save(false);
             $this->definition_id = $definition->getPrimaryKey();
-        } else if($this->isDefinitionContent() && !$definition->isNewRecord && !$definition->hasValues()) {
+        } else if($this->isDefinitionContent() && !$definition->isNewRecord && !$definition->hasValues()
+                && $this->scenario === self::SCENARIO_EDIT_ADMIN) {
             // If we reset the default definition to an empty state we remove the definition settings, which will allow content to define own definitions
            self::updateAll(['definition_id' => null], ['definition_id' => $definition->id]);
            $definition->delete();
